parameterize pbmc3k scanpy engine test
